Remove another unnecessary bridge call.
