+++ app/src/main/assets/bundle-test.js
@@ -34,70 +34,15 @@ window.onload = function() {
     module.exports.sendMessage( "DOMLoaded", {} );
 };
 },{}],2:[function(require,module,exports){
-var bridge = require( "./bridge" );
-var transformer = require("./transformer");
-
-bridge.registerListener( "setPageProtected", function( payload ) {
-    var el = document.getElementsByTagName( "html" )[0];
-    if (!el.classList.contains("page-protected") && payload.protect) {
-        el.classList.add("page-protected");
-    }
-    else if (el.classList.contains("page-protected") && !payload.protect) {
-        el.classList.remove("page-protected");
-    }
-    if (!el.classList.contains("no-editing") && payload.noedit) {
-        el.classList.add("no-editing");
-    }
-    else if (el.classList.contains("no-editing") && !payload.noedit) {
-        el.classList.remove("no-editing");
-    }
-} );
-
-bridge.registerListener( "setDecorOffset", function( payload ) {
-    transformer.setDecorOffset(payload.offset);
-} );
-
-},{"./bridge":1,"./transformer":3}],3:[function(require,module,exports){
-function Transformer() {
-}
-
-var transforms = {};
-var decorOffset = 0; // The height of the toolbar and, when translucent, status bar in CSS pixels.
-
-Transformer.prototype.register = function( transform, fun ) {
-    if ( transform in transforms ) {
-        transforms[transform].push( fun );
-    } else {
-        transforms[transform] = [ fun ];
-    }
-};
-
-Transformer.prototype.transform = function( transform, element ) {
-    var functions = transforms[transform];
-    for ( var i = 0; i < functions.length; i++ ) {
-        element = functions[i](element);
-    }
-};
-
-Transformer.prototype.getDecorOffset = function() {
-    return decorOffset;
-};
-
-Transformer.prototype.setDecorOffset = function(offset) {
-    decorOffset = offset;
-};
-
-module.exports = new Transformer();
-},{}],4:[function(require,module,exports){
 var bridge = require("../js/bridge");
 bridge.registerListener( "injectScript", function( payload ) {
     require(payload.src);
 });
-},{"../js/bridge":1}],5:[function(require,module,exports){
+},{"../js/bridge":1}],3:[function(require,module,exports){
 var bridge = require("../js/bridge");
 console.log("Something!");
 bridge.registerListener( "ping", function( payload ) {
     bridge.sendMessage( "pong", payload );
 });
 
-},{"../js/bridge":1}]},{},[2,1,4,5]);
+},{"../js/bridge":1}]},{},[1,2,3]);
