--- www/js/main.js
@@ -1,22 +1,0 @@
-var bridge = require( "./bridge" );
-var transformer = require("./transformer");
-
-bridge.registerListener( "setPageProtected", function( payload ) {
-    var el = document.getElementsByTagName( "html" )[0];
-    if (!el.classList.contains("page-protected") && payload.protect) {
-        el.classList.add("page-protected");
-    }
-    else if (el.classList.contains("page-protected") && !payload.protect) {
-        el.classList.remove("page-protected");
-    }
-    if (!el.classList.contains("no-editing") && payload.noedit) {
-        el.classList.add("no-editing");
-    }
-    else if (el.classList.contains("no-editing") && !payload.noedit) {
-        el.classList.remove("no-editing");
-    }
-} );
-
-bridge.registerListener( "setDecorOffset", function( payload ) {
-    transformer.setDecorOffset(payload.offset);
-} );
